[refactor/72-task] dto req res 구분
[refactor/72-task] dto req res 구분
