+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/application/AuthService.java
@@ -3,6 +3,11 @@ package depth.main.ideac.domain.auth.application;
 import depth.main.ideac.domain.auth.domain.Token;
 import depth.main.ideac.domain.auth.domain.repository.TokenRepository;
 import depth.main.ideac.domain.auth.dto.*;
+import depth.main.ideac.domain.auth.dto.req.FindIdReq;
+import depth.main.ideac.domain.auth.dto.req.RefreshTokenReq;
+import depth.main.ideac.domain.auth.dto.req.SignInReq;
+import depth.main.ideac.domain.auth.dto.req.SignUpReq;
+import depth.main.ideac.domain.auth.dto.res.AuthRes;
 import depth.main.ideac.domain.user.domain.Role;
 import depth.main.ideac.domain.user.domain.Status;
 import depth.main.ideac.domain.user.domain.User;
