+++ ideac-core/src/main/java/depth/main/ideac/domain/auth/dto/req/SignUpReq.java
@@ -1,4 +1,4 @@
-package depth.main.ideac.domain.auth.dto;
+package depth.main.ideac.domain.auth.dto.req;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
