+++ ideac-user/src/main/java/depth/main/ideac/domain/idea_post/dto/res/GetDetailIdeaRes.java
@@ -1,4 +1,4 @@
-package depth.main.ideac.domain.idea_post.dto;
+package depth.main.ideac.domain.idea_post.dto.res;
 
 import lombok.Builder;
 import lombok.Getter;
